BUGFIX/MINOR(grafana): Fix log-related graph title and query
 ##### SUMMARY

In OpenIO services, graphs featuring response times / bandwidth are now
marked as "gateway" when the value is proper to it. On the contrary,
response time should not be filtered, and is gathered for all services
as average.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION

Because the "max" value information is averaged across the measurement,
it is not interesting to display it. A more powerful system would be
able to gather individual max values.