CHANGE(grafana): Use openio package, fix overview dashboard
 ##### SUMMARY

On CentOS7, the package is now installed from OpenIO repositories to
better manage versions. This requires a new defaults file to be
provisioned.

The overview dashboard provisioning has also been fixed when the ansible
node cannot contact the admin node.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION

For proper package migrations, a reinstall of grafana using the OpenIO
repos is necessary