CHANGE(grafana): Use redis collector metrics
 ##### SUMMARY

Following the introduction of the redis collector, the metrics
associated to redis have changed. This ensures that the charts use new
metrics to display the same graphs. See
https://github.com/open-io/ansible-role-openio-netdata/pull/66 for more
details

Also, the uid field is persisted in the retriever tool, to stay
compatible with https://github.com/open-io/ansible-role-openio-grafana/commit/1c26344cf9399836cdce0c31eb498b6f9d6185cc. This is responsible for reorganising JSON fields on other dashboards

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION